# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fall Detection API",
        "description": "API documentation for the Fall Detection System",
        "contact": {
            "email": "admin@fall-detection.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://fall-detection-backend.test/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/elderly-profiles": {
            "get": {
                "tags": [
                    "Elderly Profiles"
                ],
                "summary": "Get all elderly profiles",
                "operationId": "f3f69917639ea6115d34bbdca639bb2a",
                "responses": {
                    "200": {
                        "description": "List of elderly profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ElderlyProfile"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Elderly Profiles"
                ],
                "summary": "Create a new elderly profile",
                "operationId": "63890e5ca4a84159fdc70d8bcdf9ea52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ElderlyProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ElderlyProfile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/elderly-profiles/{id}": {
            "get": {
                "tags": [
                    "Elderly Profiles"
                ],
                "summary": "Get elderly profile by ID",
                "operationId": "06d2c8246d545561d94c0949869795d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ElderlyProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Elderly Profiles"
                ],
                "summary": "Update elderly profile",
                "operationId": "6e36631d94441da580edd29d4ef2b034",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ElderlyProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ElderlyProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Elderly Profiles"
                ],
                "summary": "Delete elderly profile",
                "operationId": "f034885456dfc49619c17683dc9bc164",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile deleted successfully"
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ElderlyProfile": {
                "required": [
                    "first_name",
                    "last_name",
                    "date_of_birth",
                    "gender",
                    "primary_phone",
                    "current_address",
                    "emergency_contact_name",
                    "emergency_contact_phone",
                    "emergency_contact_relationship",
                    "mobility_status",
                    "vision_status",
                    "hearing_status",
                    "primary_carer_id",
                    "care_level",
                    "preferred_language",
                    "device_status",
                    "living_situation",
                    "activity_level"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1940-01-01"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "other"
                        ],
                        "example": "male"
                    },
                    "profile_photo": {
                        "type": "string",
                        "example": "profile-photos/john.jpg",
                        "nullable": true
                    },
                    "height": {
                        "type": "number",
                        "format": "float",
                        "example": 170.5,
                        "nullable": true
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 65.2,
                        "nullable": true
                    },
                    "blood_type": {
                        "type": "string",
                        "example": "O+",
                        "nullable": true
                    },
                    "national_id": {
                        "type": "string",
                        "example": "A123456789",
                        "nullable": true
                    },
                    "primary_phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "secondary_phone": {
                        "type": "string",
                        "example": "+0987654321",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com",
                        "nullable": true
                    },
                    "current_address": {
                        "type": "string",
                        "example": "123 Main St, City, Country"
                    },
                    "emergency_contact_name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "emergency_contact_phone": {
                        "type": "string",
                        "example": "+1122334455"
                    },
                    "emergency_contact_relationship": {
                        "type": "string",
                        "example": "Daughter"
                    },
                    "medical_conditions": {
                        "type": "string",
                        "example": "Diabetes",
                        "nullable": true
                    },
                    "allergies": {
                        "type": "string",
                        "example": "Peanuts",
                        "nullable": true
                    },
                    "current_medications": {
                        "type": "string",
                        "example": "Metformin",
                        "nullable": true
                    },
                    "disabilities": {
                        "type": "string",
                        "example": "Hearing loss",
                        "nullable": true
                    },
                    "mobility_status": {
                        "type": "string",
                        "enum": [
                            "independent",
                            "needs_assistance",
                            "wheelchair_bound"
                        ],
                        "example": "independent"
                    },
                    "vision_status": {
                        "type": "string",
                        "enum": [
                            "normal",
                            "glasses",
                            "impaired"
                        ],
                        "example": "normal"
                    },
                    "hearing_status": {
                        "type": "string",
                        "enum": [
                            "normal",
                            "hearing_aid",
                            "impaired"
                        ],
                        "example": "normal"
                    },
                    "last_medical_checkup": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-12-01",
                        "nullable": true
                    },
                    "primary_carer_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "secondary_carer_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "care_level": {
                        "type": "string",
                        "enum": [
                            "basic",
                            "moderate",
                            "intensive"
                        ],
                        "example": "basic"
                    },
                    "special_care_instructions": {
                        "type": "string",
                        "example": "Needs help with medication",
                        "nullable": true
                    },
                    "daily_routine_notes": {
                        "type": "string",
                        "example": "Walks every morning",
                        "nullable": true
                    },
                    "dietary_restrictions": {
                        "type": "string",
                        "example": "No sugar",
                        "nullable": true
                    },
                    "preferred_language": {
                        "type": "string",
                        "example": "English"
                    },
                    "device_id": {
                        "type": "string",
                        "example": "DEV123456",
                        "nullable": true
                    },
                    "device_status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "last_device_check": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-05-01",
                        "nullable": true
                    },
                    "device_battery_level": {
                        "type": "integer",
                        "example": 85,
                        "nullable": true
                    },
                    "device_location": {
                        "type": "string",
                        "example": "Home",
                        "nullable": true
                    },
                    "preferred_hospital": {
                        "type": "string",
                        "example": "City Hospital",
                        "nullable": true
                    },
                    "insurance_information": {
                        "type": "string",
                        "example": "Provider: ABC Insurance, Policy: 12345",
                        "nullable": true
                    },
                    "living_situation": {
                        "type": "string",
                        "enum": [
                            "lives_alone",
                            "with_family",
                            "assisted_living"
                        ],
                        "example": "with_family"
                    },
                    "activity_level": {
                        "type": "string",
                        "enum": [
                            "active",
                            "moderate",
                            "sedentary"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Prefers vegetarian meals",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "carer",
                            "elderly"
                        ],
                        "example": "elderly"
                    },
                    "device_token": {
                        "type": "string",
                        "example": "fcm_token_here",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "elderly_profile": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/ElderlyProfile"
                            }
                        ],
                        "nullable": true
                    },
                    "carer_profile": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CarerProfile"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}